pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dora-amiga"
version = "0.0.1"
authors = [{ name = "Edgar Riba", email = "edgar@farm-ng.com" }]
description = "Dora bridge for the Farm-ng Amiga"
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "dora-rs >= 0.3.9",
    "farm-ng-amiga",
    "kornia_rs",
    "opencv-python",
    "numpy",
    "foxglove-sdk"
]

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
dora-amiga-camera = "dora_amiga.camera:main"
dora-amiga-gps = "dora_amiga.gps:main"
dora-amiga-canbus = "dora_amiga.canbus:main"
dora-amiga-imu = "dora_amiga.imu:main"
dora-amiga-foxglove = "dora_amiga.foxglove:main"

[tool.ruff.lint]
extend-select = [
  "D",    # pydocstyle
  "UP",   # Ruff's UP rule
  "PERF", # Ruff's PERF rule
  "RET",  # Ruff's RET rule
  "RSE",  # Ruff's RSE rule
  "NPY",  # Ruff's NPY rule
  "N",    # Ruff's N rule
  "I",    # Ruff's I rule
]
